Adds threading for messaging outside main loop, fixes msg function's default target logic
Unsure at the moment how to structure the threading and potential future functions, I'd rather have the scheduled message thread be in its separate file to unclutter the main python file. But the threaded function requires the same instance of the instantiated IRC class. The logic that the thread should handle could perhaps be handled by another class.